feat: export feature - fix cancellation bugs - fix chrome download bugs - refactor export component : extract
